Ambient Display: Include a custom setting to stop android from inverting the colors of the notifications on Ambient Display - Part 1
Thanks @dankoman30 for the original idea. Given CMTE doesn't allow for
bool values to be changed, we have implemented a switch in order to
implement this functionality.
This is part one of the commit, and it will implement the necessary
changes on framework and SystemUI. The switch will be included in part
2.
NOTE: for now the default is set to “false” which means that the
notifications will NOT be inverted unless the switch changes that.